+++ core/src/db/DBImpl.cpp
@@ -171,6 +171,8 @@ DBImpl::Stop() {
             bg_flush_thread_.join();
         }
 
+        WaitMergeFileFinish();
+
         swn_index_.Notify();
         bg_index_thread_.join();
 
@@ -1262,22 +1264,22 @@ DBImpl::BackgroundIndexThread() {
 
 void
 DBImpl::WaitMergeFileFinish() {
-    ENGINE_LOG_DEBUG << "Begin WaitMergeFileFinish";
+    //    ENGINE_LOG_DEBUG << "Begin WaitMergeFileFinish";
     std::lock_guard<std::mutex> lck(merge_result_mutex_);
     for (auto& iter : merge_thread_results_) {
         iter.wait();
     }
-    ENGINE_LOG_DEBUG << "End WaitMergeFileFinish";
+    //    ENGINE_LOG_DEBUG << "End WaitMergeFileFinish";
 }
 
 void
 DBImpl::WaitBuildIndexFinish() {
-    ENGINE_LOG_DEBUG << "Begin WaitBuildIndexFinish";
+    //    ENGINE_LOG_DEBUG << "Begin WaitBuildIndexFinish";
     std::lock_guard<std::mutex> lck(index_result_mutex_);
     for (auto& iter : index_thread_results_) {
         iter.wait();
     }
-    ENGINE_LOG_DEBUG << "End WaitBuildIndexFinish";
+    //    ENGINE_LOG_DEBUG << "End WaitBuildIndexFinish";
 }
 
 void